refactor: 🎨 mv clean syntax context to visitors folder
